backend/web/requirements.txt
google-cloud-storage
flask
pytest
firebase_admin
